+++ Assets/Prefabs/Graph/DecryptionNode.prefab
@@ -190,7 +190,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1308720565020048102}
   serializedVersion: 2
-  m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0.78, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
@@ -316,7 +316,7 @@ MonoBehaviour:
   nodeState: {fileID: 0}
   levelManager: {fileID: 0}
   nodeRenderer: {fileID: 0}
-  minigameSceneName: 
+  minigameSceneName: Pattern-Matching
 --- !u!114 &1483371888085921634
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -331,8 +331,8 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   node:
     nodeType: 6
-    visible: 1
-    compromised: 1
+    visible: 0
+    compromised: 0
     initCompromisationTime: 0
 --- !u!114 &3674063514402068433
 MonoBehaviour:
