+++ Assets/Prefabs/UIElements/MinigameCanvas.prefab
@@ -17,7 +17,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!224 &5044784538391944262
 RectTransform:
   m_ObjectHideFlags: 0
