+++ Assets/Scenes/Minigames/Fire Pacman.unity
@@ -318,20 +318,6 @@ GameObject:
   m_CorrespondingSourceObject: {fileID: 5874001750792411898, guid: 2231f643ccff22d4998c5c85f3fa304e, type: 3}
   m_PrefabInstance: {fileID: 1106765389}
   m_PrefabAsset: {fileID: 0}
---- !u!114 &335050060
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 335050059}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 0c56d20322766e54dae048451531107d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  hackerman: {fileID: 1197567267}
-  agent: {fileID: 0}
 --- !u!50 &335050065
 Rigidbody2D:
   serializedVersion: 4
@@ -2677,7 +2663,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 5874001750792411898, guid: 2231f643ccff22d4998c5c85f3fa304e, type: 3}
       propertyPath: m_IsActive
-      value: 0
+      value: 1
       objectReference: {fileID: 0}
     - target: {fileID: 6588899654408981625, guid: 2231f643ccff22d4998c5c85f3fa304e, type: 3}
       propertyPath: m_Radius
@@ -3123,8 +3109,12 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 5874001750792411898, guid: 2231f643ccff22d4998c5c85f3fa304e, type: 3}
       propertyPath: m_IsActive
-      value: 0
+      value: 1
       objectReference: {fileID: 0}
+    - target: {fileID: 6568652077087580880, guid: 2231f643ccff22d4998c5c85f3fa304e, type: 3}
+      propertyPath: hackerman
+      value: 
+      objectReference: {fileID: 1197567267}
     - target: {fileID: 6588899654408981625, guid: 2231f643ccff22d4998c5c85f3fa304e, type: 3}
       propertyPath: m_Radius
       value: 0.25
@@ -3192,9 +3182,6 @@ PrefabInstance:
     m_RemovedGameObjects: []
     m_AddedGameObjects: []
     m_AddedComponents:
-    - targetCorrespondingSourceObject: {fileID: 5874001750792411898, guid: 2231f643ccff22d4998c5c85f3fa304e, type: 3}
-      insertIndex: -1
-      addedObject: {fileID: 335050060}
     - targetCorrespondingSourceObject: {fileID: 5874001750792411898, guid: 2231f643ccff22d4998c5c85f3fa304e, type: 3}
       insertIndex: -1
       addedObject: {fileID: 335050065}
@@ -3320,7 +3307,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 5874001750792411898, guid: 2231f643ccff22d4998c5c85f3fa304e, type: 3}
       propertyPath: m_IsActive
-      value: 0
+      value: 1
       objectReference: {fileID: 0}
     - target: {fileID: 6588899654408981625, guid: 2231f643ccff22d4998c5c85f3fa304e, type: 3}
       propertyPath: m_Radius
@@ -3386,6 +3373,7 @@ PrefabInstance:
     - {fileID: 2218877854671218792, guid: 2231f643ccff22d4998c5c85f3fa304e, type: 3}
     - {fileID: 6739350631407235764, guid: 2231f643ccff22d4998c5c85f3fa304e, type: 3}
     - {fileID: 4806701017408000381, guid: 2231f643ccff22d4998c5c85f3fa304e, type: 3}
+    - {fileID: 6568652077087580880, guid: 2231f643ccff22d4998c5c85f3fa304e, type: 3}
     m_RemovedGameObjects: []
     m_AddedGameObjects: []
     m_AddedComponents:
@@ -11315,8 +11303,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 0c56d20322766e54dae048451531107d, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  hackerman: {fileID: 1197567267}
-  agent: {fileID: 0}
 --- !u!50 &1841505078
 Rigidbody2D:
   serializedVersion: 4
@@ -11361,8 +11347,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 0c56d20322766e54dae048451531107d, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  hackerman: {fileID: 1197567267}
-  agent: {fileID: 0}
 --- !u!50 &1845895190
 Rigidbody2D:
   serializedVersion: 4
@@ -11584,11 +11568,11 @@ PrefabInstance:
     m_Modifications:
     - target: {fileID: 4322814095124702754, guid: a03495a331c9f8c4980b82c01bcd6d5a, type: 3}
       propertyPath: m_LocalPosition.x
-      value: -37.5
+      value: -37.461
       objectReference: {fileID: 0}
     - target: {fileID: 4322814095124702754, guid: a03495a331c9f8c4980b82c01bcd6d5a, type: 3}
       propertyPath: m_LocalPosition.y
-      value: 16.5
+      value: 16.39
       objectReference: {fileID: 0}
     - target: {fileID: 4322814095124702754, guid: a03495a331c9f8c4980b82c01bcd6d5a, type: 3}
       propertyPath: m_LocalPosition.z
@@ -11622,34 +11606,14 @@ PrefabInstance:
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 5223019629174184618, guid: a03495a331c9f8c4980b82c01bcd6d5a, type: 3}
-      propertyPath: hackerman
-      value: 
-      objectReference: {fileID: 1197567267}
     - target: {fileID: 6018925601545237779, guid: a03495a331c9f8c4980b82c01bcd6d5a, type: 3}
       propertyPath: m_Name
       value: NavFlame
       objectReference: {fileID: 0}
-    - target: {fileID: 6018925601545237779, guid: a03495a331c9f8c4980b82c01bcd6d5a, type: 3}
-      propertyPath: m_IsActive
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6445593185529324432, guid: a03495a331c9f8c4980b82c01bcd6d5a, type: 3}
-      propertyPath: m_Radius
-      value: 0.25
-      objectReference: {fileID: 0}
-    - target: {fileID: 8350835390970777677, guid: a03495a331c9f8c4980b82c01bcd6d5a, type: 3}
-      propertyPath: m_Color.b
-      value: 0.06738185
-      objectReference: {fileID: 0}
-    - target: {fileID: 8350835390970777677, guid: a03495a331c9f8c4980b82c01bcd6d5a, type: 3}
-      propertyPath: m_Color.g
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8350835390970777677, guid: a03495a331c9f8c4980b82c01bcd6d5a, type: 3}
-      propertyPath: m_Color.r
-      value: 0.5377358
-      objectReference: {fileID: 0}
+    - target: {fileID: 9183313837885946770, guid: a03495a331c9f8c4980b82c01bcd6d5a, type: 3}
+      propertyPath: hackerman
+      value: 
+      objectReference: {fileID: 1197567267}
     m_RemovedComponents: []
     m_RemovedGameObjects: []
     m_AddedGameObjects: []
