+++ Assets/Scripts/Systems/Graph/BasicNodeController.cs
@@ -4,7 +4,6 @@ public class BasicNodeController : NodeController
 {
     public override void OnMouseDown()
     {
-        Debug.Log("Clicked");
         levelManager.AddNodeToDataPath(nodeState);
     }
