+++ Assets/Scripts/Systems/Minigames/Fire Pacman/Flame.cs
@@ -1,24 +1,58 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.AI;
 
 public class Flame : MonoBehaviour
 {
     public Movement movement { get; private set; }
-    public FlameChase chase { get; private set; }
-    public Transform hackerman;
+    private Transform hackerman;
+    private NavMeshAgent agent;
+
+    private FirewallMinigameManager firewallMinigameManager;
+
     public void Awake()
     {
         this.movement = GetComponent<Movement>();
-        this.chase = GetComponent<FlameChase>();
+        this.agent = GetComponent<NavMeshAgent>();
+        this.hackerman = FindAnyObjectByType<Hackerman>().transform;
+        agent.updateRotation = false;
+        agent.updateUpAxis = false;
+        agent.SetDestination(hackerman.position);
+        //InvokeRepeating("SetDestination", 2f, 1f);
+        firewallMinigameManager = FindAnyObjectByType<FirewallMinigameManager>();
     }
 
     private void OnCollisionEnter2D(Collision2D collision)
     {
+        agent.SetDestination(hackerman.position);
         if (collision.gameObject.layer == LayerMask.NameToLayer("HackerMan"))
         {
-            FindObjectOfType<FirewallMinigameManager>().GameOver(false);
+            FindAnyObjectByType<FirewallMinigameManager>().GameOver(false);
+        }
+    }
+
+    private void OnTriggerEnter2D(Collider2D collision)
+    {
+        Path path = collision.GetComponent<Path>();
+
+        if (path != null)
+        {
+            agent.SetDestination(hackerman.position);
+        }
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        if (firewallMinigameManager.MinigameInProgess)
+        {
+            agent.isStopped = false;
+            if (agent.remainingDistance < 0.3)
+            {
+                agent.SetDestination(hackerman.position);
+            }
+        } else
+        {
+            agent.isStopped = true;
         }
     }
 }
