+++ Assets/Scripts/Systems/Minigames/Fire Pacman/Flame.cs.meta
@@ -1,2 +1,2 @@
 fileFormatVersion: 2
-guid: 8c3bc2989d5c6bc40a032e3cc9e55637
+guid: 0c56d20322766e54dae048451531107d
