--- Assets/Scripts/Systems/Minigames/Fire Pacman/Flame1.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 0c56d20322766e54dae048451531107d
