--- Assets/Scripts/Systems/Minigames/Fire Pacman/FlameChase.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 546dc43b5af14c24793528e0e0266539
